Updates to FT and force_single
Bug fixes and updates. ft2d and ft3d were removed since they are part of matlab_tools funcitoning independent of CudaMat.

+++ single_force.m
@@ -22,8 +22,8 @@
 
 function out = single_force(in)
 if isa(in,'cuda')
-    out = double(cuda_cuda('get',getReference(in)));
+    out = single(cuda_cuda('get',getReference(in)));
 else
-    out=double(in);
+    out=single(in);
 end
